requirement.txt
torch
PyYAML
timm
tqdm
matplotlib
accelerate
evaluate
albumentations
torchinfo
ptflops